Brunel runs

+++ libnestutil/nest_types.h
@@ -129,7 +129,7 @@ __attribute__( ( __unused__ ) ) constexpr size_t invalid_index = std::numeric_li
 /**
  *  For enumerations of synapse types.
  */
-typedef unsigned int synindex;
+typedef size_t synindex; // PYNEST-NG check if size_t instead of unsigned int causes problems
 const synindex invalid_synindex = MAX_SYN_ID;
 
 /**
